add support for self-intersecting torus.
this extends support for defining and editing torus so that the minor radius (r_h aka r2) can be bigger than the major radius (r_a aka r1).  intersection logic is updated to report hits correctly (i.e., only one segment through the overlapped interior).  leaving plotting as-is since it is useful for the user to have some visual cue.

this helps improve usability a smidge so objects can be edited through a larger spectrum of valid values, and supports definition of some shapes that are otherwise impossible with the implicits.