massive overhaul of rtwizards option printing.
turns out, there was a bug where -h simply displayed all options in triplicate.  now correctly filters to a subset, with subsets reorganized and ordered according to functionality, priority, and simplicity of presentation.  Should be a lot easier for a passing user to set up a valid rendering via command line.