add a script that characterizes a surface area metric based off of presented area and edge rendering information.
the script uses rtedge, rtarea, rtwizard, pixcount, and imagemagick to compose an image with metrics that characterize external surface area complexity.  this complexity metric can be views as a proxy for surface area.  using edge information, and some baked-in notions of time/effort, we could even give a rough order of magnitude effort estimate for what it'd take to go over the surface (e.g., to paint it,  wash it, apply coatings, examine, etc).  everything is generated automatically and composited (with labels) via  imagemagick tools.