Add support for prefixing to plot3.h
In order to allow libbg to use plot3.h independently of libbv and not cause
problems when static linking with other libs, use the same trick we use for
zlib and allow for a custom prefix.  In the libbg uses of plot3 we define a
separate libbg specific prefix, and for all other code just use the default
behavior.  This should allow libbg to define PLOT3_IMPLEMENTATION without
causing conflicts with libbv's use of PLOT3_IMPLEMENTATION, but needs to be
tested.