fix rtweight printing inside mged
rtweight was only printing stderr output when run inside mged despite
writing to both stderr and stdout - add an io hook for stdout. Also,
since _ged_rt_output_handler() is setup to handle BOTH stderr and
stdout, we need to restructure slightly so that successful reads are
printed to the console before the next overwrites the buffer.