Better explain the implications of face_mode BoTs
The raytacing behavior of plate mode bots using face_mode is view
dependent.  Because what volume is claimed as part of the "solid"
changes based on the incoming ray direction, what we end up doing for
those cases is making a solid BoT that claims all the volume potentially
claimed.

For any given ray, this means the shotline's solid segment will be
longer, since the solid volume will end up reporting both the "before"
and "after" hit segments from the perspective of the original plate mode
BoT.  There isn't a perfect answer here since view dependent behavior
isn't something we can represent with a volumetric BoT.

If the user wants to create a version of the bot that will return
similar shotline lengths, they can adjust the per-face thickness
to half that of its current value to produce a smaller solid volume.