Checkpoint reworking of facetize option handling
Looking at the region code, and the prospect of adding more tessellation
methods to the backend, it becomes clear that we need a better way to
pass method specific options from the facetize command line rather than
hard-coding the specific parameters of each method at the top level.

Also, since we're using a working copy of the .g now to do the
processing, the logic for managing all the region copies may be
unnecessary.  A combination of keep and dbconcat using the working .g as
a starting point should be able to achieve the same thing we're trying
to achieve with all the comb copying in the current code, so let's see
if we can simplify.