Add -F opt to extrude for view dependent inputs
After discussion with Sean, default the bot extrude behavior to
rejecting view dependent inputs unless the user specifically adds a flag
to proceed with them anyway.  The intent here is to force users to
recognize that their original geometry cannot be accurately reproduced
as a volumetric primitive, and require them to make an informed decision
as to whether the changes implied by switching in a volumetric extrusion
are acceptable for the specific use case in question.