removed the non-merzenne twister spherical sampling during ao rendering.
while slower, there's not a use case where you want ao but would be okay with the pattern artifacts from the fast random number generator.  this change leaves the ambSlow variable to allow time for scripts to transition without hitting an error, but it's a no-op.  simplifies usability and docs.