Use empty BoTs for empty-but-successful boolean evaluation scenarios.
This gives us a way to distinguish between "we didn't get anything because the
conversion failed" and "we didn't get anything because the boolean CSG
evaluated to no volume".  conversion.sh will now be able to recognize that
we're handling things like unioned half primitives, as well as not failing in
cases where a region really does evaluate to empty volume due to subtractions
or whatnot.