Make the dbconcat man page easier to use.
Add an options table, adjust the explanation of what the implications of
affix, suffix and straight import are.

The behavior of this command needs to be improved - we should be
capable of maintaining the integrity of imported combs through renaming
automatically, rather than requiring hunting around with dup to find
a non-conflicting affix string for the purpose.