Rework dbconcat for less aggressive prefix use
Want behavior where a user supplied prefix/suffix is used ONLY when
necessary to resolve a name conflict - in other words, keep as many of
the original names as possible.

Old behavior - apply it to everything if specified - is available with
the addition of a -A option.

Update the man page - I think it is outdated even with main's dbconcat
in that combs do seem to get the word if object names need to be changed
on import.

This needs testing - it's a significant rework, using bu_vls_incr rather
than having one-off code in concat for incrementing the names, using
bu_opt, storing more info in the state struct rather than passing around
lots of parameters.