Teach bot extrude to produce volumetric BoT output
Rather than the large CSG tree originally produced by bot extrude, take
advantage of the new Manifold library's capabilities to make an
explicitly evaluated volumetric BoT by doing facetization of the
individual CSG elements and then evaluating the unions.

This can be quite slow - the boolean problem is a difficult one, with
thousands of elements even for modestly sized meshes - so users should
allow for that when preparing to convert large numbers of plate mode
solids.  For example, the FAA's Generic_Twin example in BRL-CAD takes
about an hour to complete a conversion of all its plate mode BoTs to
volumetric BoTs.

Bringing this work over from the facetize branch, since it doesn't
require Geogram or any of the new facetize logic itself.