make regress repository test fail even harder.
it started failing recently because we weren't previously matching on #elif lines and a reordering in libbu made a platform symbol suddenly be seen.  this should catch more of the possible conditions, so our counts may go up further still.  this should be a more accurate (and now documented) match now catching more C/C++ and CMake cases.