Test a method for bot vertex fusing
The Asset Import "x" regression test is producing an "arb4" bot that is
plate mode and not sharing vertices.  Intuitively we ought to be able to
"fuse" this into a proper arb4 solid.

This can be done with the sequence:

bot set type unnamed.s solid
bot_fuse tetra.s unnamed.s

but that logic goes through libnmg and it's not clear how well it would
scale.  I was curious to see if a simple vertex binning approach might
be able to achieve a similar result, which is what this code implements.
It's a "quick and dirty" test so far, so not announcing as a user
feature for now.