Update behavior of dbconcat command's prefix/suffix application
Previously, when the user suffied a prefix or suffix string to dbconcat,
it would always apply the prefix and suffix to all incoming object
names, regardless of whether the incoming name conflicted with any
object already in the database.  While there are potential scenarios
where this property is useful - it is one potential way to identify
objects that came from the concated database - it is a hinderance when
the goal is to merge the incoming geometry with a minimum of renaming
but using less generic alteration strings than the default no
prefix/suffix dbconcat naming behavior.

What we do if a user prefix/suffix is specified by default is to only
use it for renaming if we need to in order to resolve a naming conflict.
If the user wishes to have the previous "always applied" behavior, they
can get it by adding the -A option flag along with the prefix/suffix
option.

The initial motivation for this is the use of the new facetize command
implementation's region mode, which operates on a working copy and then
dbconcats the result back into the original file.  We want to prefix the
copies of the top level combs, but not (unless necessary) the new BoT
solid primitives, which usually will have no corresponding conflicting
object in the original file.