Prototype a more modularized build approach
Instead of the three levels previously defined, explore supporting
specifying one or more "components" of BRL-CAD to be enabled.  The idea
would be that if a user defines (say) -DBRLCAD_COMPONENTS=librt then
only librt and its dependencies would be enabled for compilation.

For this to be properly realized a lot of higher level build logic will
have to be conditionalized based on whether various targets are
available (for example, regression tests will need to be disabled if
their executables aren't defined as targets) but this should let us be
way more flexible when it comes to "only build what we want" situations.