Test mechanism to optionally require targets to build man pages
This will need some thought, since we originally consolidated the man
page targets to avoid exploding the target count in Visual Studio, but
to support a cleaner modular build we need to avoid building man pages
or things that have been disabled.

One possibility might be to handle the Docbook building right in the
BRLCAD_ADDEXEC and/or plugin functions and make them part of the main
exec/plugin build.