Do header sorting in per-target assignment setup
We don't want to give up the preferential ordering we had established
for our local headers with the previous include scheme, so rather than
using generator expressions (which offer no opportunity to arrange
the order of accumulated headers) we read the target properties and
build up include lists that can be sorted at configure time.

Probably worth examining more closely which headers end up in public and
private groupings - not sure manifold and poly2tri need to be in public,
for example - but for now this builds.

TODO: Individual additions to the exec includes probably should be
specifying the BEFORE property to make sure they receive preferential
ordering, unless we're locally adding a system find_package result...