With the new bext logic, modularlize/localize find
misc/CMake/BRLCAD_ExternalDeps.cmake holds the find_package logic for
the bext components, but for a modular build we don't want to be doing
all the find_package calls right away.  If some components are disabled
we never need to look for their dependencies at all.

To make this work, we define macros that encapsulate what is needed for
a resettable find process for each package (we use macros rather than
functions deliberately here, otherwise the function scope encapsulates
the find_package output variables and we would have to PARENT_SCOPE
export the ones we care about - that's possible, but for now it's extra
work and we can do it later if it looks like it would be beneficial.)

The new macros are then called in the subdirectories where those
packages are actually needed/used.  Because the scope of the
find_package results isn't global except for variables being written to
the cache, we call the macros separately in the various places we need
them, but that should be acceptable - the results should always be the
same, and the extra work of repeating the finds is minimal compared to
the overall build and modularity benefits.

This lets us build (for example) librt and its dependencies with:

cmake .. -DBRLCAD_COMPONENTS=librt && make -j3

If we let the BRL-CAD cmake configure also do its own bext build, it
will try to enable only the bext dependencies needed for the active
configuration (a pre-supplied bext, on the other hand, is what it is -
the logic will not try to pick and choose just the files needed from an
already created output directory, as there is no way to reliably tell
what is needed and what isn't.)

The next step is probably to look at whether it makes sense to push the
various configure checks for headers and functions down closer to where
they are needed as well.  That would probably mean revisiting
brlcad_config.h - we'd either need to define private headers for the
various individual libs and programs to use instead, or supply command
line defines (the latter might make more sense if we only have a small
number of defines relevant to a particular component.)  Would also need
to determine which parts of brlcad_config.h really need to be part of
the install, and which parts are build-time only.