Stash an rle version of the Tk test teapot img
It's looking as if we didn't hook UtahRLE conversion support into
libicv, which means we don't currently have the ability to read and
write them exposed in our tools - need to fix that.

rle images aren't common, so we create one by using the netpbm utility
pnmtorle to convert the Tk test image of the teapot, and then passed
that result through the old Utahrle tools rletoraw and rawtorle.  Using
the netpbm rletopnm to convert the result back into something viewable
indicates the conversions were sane, so this should at least give us
something to work with.  Also add the original Tk version of the image
(which is already in our repo history anyway via Tk) to use as a
control.  In principle we ought to be able to read these both in with
libicv and get a clean diff comparison.