Consoliate find_package resets into BRLCAD_Find_Package
Per suggestion from Sean, most (but not all) of the find_package cases respect
a similar pattern.  Rather than have one macro per package, it would be better
to make a generic wrapper that handles most of the cases and use it.

The key to making this work is the CACHE_VARIABLES directory property, which
gives us a way to get a "before and after" insight into what a find_package
call has done to the cache.  Once we know that, we can use unset(CACHE) to
clear those values from the cache.  We also make sure to save and then restore
the values to the working variable state, so we don't destroy the find_package
results for the working session.

We do, however, still need something in the cache so the BRL-CAD Summary
printing knows what happened - the find_package variables are scoped in their
calling directories and aren't directly visible.  To handle this, we write out
a *_STATUS variable for each package and teach the Summary logic to look for
it.  The *_STATUS variable isn't functional, so it is harmless as far as
find_package is concerned, but it still gives us what we need for the final
summary.  Nor does it matter that the value will be overwritten by mulitple
calls, since the results of multiple BRLCAD_Find_Package calls with the same
package should always be the same.  (If not, it is an error.)