Remove the src/other and misc/tools/ext dirs
The motivation for doing this is to fully decouple our third party
bulding logic from BRL-CAD's primary build and allow our use of third
party components to scale beyond just what is feasible within our main
tree.  Integration of independent build systems has always been a source
of issues, and when it comes to very large prospective dependencies
(Boost, LLVM, Qt, etc.) our existing approaches simply weren't
practical.  Consequently, a new appoach was developed - using a
completely separate repository to focus solely on the management of
third party components.  In this new scheme the only role BRL-CAD's
build system has is to copy and bundle provided third party outputs, as
well as pointing CMake's "find_package" logic preferentially to those
local copies if they are present.  The new "gist" tool already depends
on this approach when bundling OpenCV; that dependency was never
integrated into src/other/ext, and so requires either a system version
or a bext-based build.  There is also additional experimental code in
branches and off-by-default compilation units that already depends on
bext.

In bext-based compilation, if user does nothing to specify any bext
inputs the BRL-CAD configure process will attempt to do its own clone,
configure, and build of the elements of bext necessary for the current
build configuration, based on the configuration settings.  If a
pre-existing copy of the bext source tree is present in an expected
location, the clone step can be skipped - otherwise a network
connection must be present to download the sources.  If BRL-CAD's
configure is doing the configure+build steps for bext, it will enable
only the necessary compoennts.

If the user has a pre-compiled bext output directory they wish the build
to use, rather than building components as part of configure, the
BRLCAD_EXT_DIR variable may be used to point the build to that tree.
Note that in that mode, BRL-CAD's build will bundled whatever files are
present in the provided outputs - for example, if Qt6 is present in the
bext output directory specified but BRL-CAD's BRLCAD_ENABLE_QT feature
is off, a generated output package will still contain Qt6 even though it
is unused per the BRL-CAD configure and build settings.

We have been working to shift our CMake build logic from in-tree
src/other and misc/tools/ext directories to pulling from the
https://github.com/BRL-CAD/bext respository.  Logic is already in place
to support the latter, conditionalized on whether src/other and
misc/tools/ext are present in the source tree.

By removing these two directories, we are making the bext logic "live" -
testing has gone well thus far, so it is time to iron out any remaining
practical issues with this approach.