Add ifdef guarded tessellation code using detria
The detria header-only implementation of CDT is looking very
interesting.  Upstream changes made it work for a C++17 build, so add it
to the build for more extensive testing, guarded by a USE_DETRIA ifdef
check.

Have tried a few simple cases - facetize of a brep rcc, shaded drawing
of a few NURBS brep objects - but will need *extensive* testing to
verify it works as we need it to, which is why it's added behind an
ifdef conditional.  If this does pan out however, it would allow us to
remove our older, semi-forked poly2tri version in favor of a new
header-only implementation that is much simpler to integrate.  Upstream
testing results suggest it may also be somewhat faster.