Set up bext as a submodule in src
A default "git clone" of BRL-CAD will not populate bext, so the more
minimal configure-driven "clone only what you need" process will still
be the default.  However, since that requires an internet connection to
work, we also want to make it as easy as possible to get a "full"
checkout of all the necessary sources, and setting up a submodule will
allow for the possibility of a "one shot clone everything" process with:

    git clone --recursive https://github.com/BRL-CAD/brlcad.git

This will result in a large checkout, since all the Appleseed and OSPray
dependencies will also be cloned, but it will obtain everything with one
command.