Check to see if a populated src/bext is up to date
Because git submodules always point to a specific SHA1, there's no way
to tell src/bext to always pull the latest checkout of RELEASE when
doing the "git clone --recurse-submodules" checkout.  Since bext and
brlcad are decoupled, updating BRL-CAD/bext's RELEASE branch won't
automatically update the src/bext submodule.  In order to keep the two
in sync, in the cases where we have an internet connection validate the
local submodule SHA1 against the bext RELEASE branch's latest SHA1.
It's not a fatal error - we just warn if there's a mismatch and advise
updating src/bext to point to the latest.

In the case of a checkout of a release tag of BRL-CAD, src/bext should
be pointed to the equivalent rel-X-Y-Z branch of bext rather than
RELEASE.  Accordingly, we also check for that case if the RELEASE SHA1
doesn't match before concluding there's a problem.