Add bot repair command to expose Geogram hole filling
The "repair" subcommand to the libged "bot" command will check if a mesh
is manifold.  If not, it tries to use the Geogram library's hole filling
routines to produce a manifold mesh.  It can either write out the new
mesh in-place, or to a user-supplied object name.

Note that this is intended to repair relatively small holes in nearly
manifold meshes - if a mesh is sufficiently messy, the original modeling
intent is lost and even a "repaired" mesh may not be very representative
of what an actual human modeler would have reconstructed.