Expose the Manifold check as a bot check cmd
Manifold library's validity testing is key for determining whether a
given BoT is viable for boolean evaluation, so wire it in to the bot
check command.  It's added to the default checks and "solid" (solid is
the test implicitly selected when no specific test is asked for.)  The
specific Manifold test can also be run in isolation with the "manifold"
subcommand:

   bot check manifold input.bot

A warning will be printed if BRL-CAD's routines and Manifold disagree -
such a case probably warrants investigation to see if there is a
problem in either testing setup.