bot extrude enhancement to produce an evaluated BoT
Taking advantage of the Manifold boolean abilities, we are able to
properly realize the "bot extrude" feature and produce a mesh that is
the result of the evaluated CSG tree of primitives generated to
represent the implicit volume of a plate mode BoT explicitly.  The comb
tree that was previously produced can still be generated with the -C
option, but this is not recommended as those comb trees can be very
large and difficult to work with.

For large plate mode BoTs the extrusion process may be rather slow - the
boolean problem involved with evaluating the union of all individual
elements of the extrusion is difficult - but the output should be a
properly manifold representation of the plate mode volume.

Note that the output volume is ultimately an approximation of the
implicit plate mode volume - even in the simplest cases there can be
view dependent behavior of plate mode BoTs depending on incoming ray
angles.  Plate modes where the hit distance is always extruded out in
the direction of the ray are ignored - such behaviors are not
representable with volumetric geometry, and the standard "extrude in
both directions from the middle" volume will be generated regardless.