Force all CMake generators to use one configuration
With the switch to bext, the configure process (in some modes) needs to
build bext itself.  This means with MSVC we can't switch the build type
from Debug to Release at build time, since Debug bext binaries aren't
usable with Release BRL-CAD build outputs and vice versa.  Even with a
pre-defined BRLCAD_EXT_DIR, that directory will hold only one
configuration of the bext binaries and we still won't be able to do
a Debug/Release build time switch.

If we wanted the configure process in a multiconfig build environment to
pre-build and manage all possible configurations of bext (which can be
as many as 4, if we use the CMake pre-defined build configurations) we
would have a huge problem in terms of build logic complexity, disk space
and configure time requirements as we built multiple copies of
potentially all of bext.  This isn't worthwhile just to be able to
select between Debug and Release at built time, when we can just set up
two build directories to achieve the same result and avoid the whole
problem.

To constrain BRL-CAD's build to the desired behavior, we need to
simplify what configurations are allowed for a single configure+build
setup by overriding CMAKE_CONFIGURATION_TYPES.  As a bonus, we can also
eliminate a lot of logic that was added specifically to support BRL-CAD
being runnable in per-configuration build directories.  Needs to be
tested with Visual Studio to make sure we've successfully cleared the
old logic out correctly and set up the necessary output directory
variable overrides to make things work like a single-config tool, but
in the end this should be a lot easier to deal with across the board.