Rough logic to spot between bext and CMAKE_BUILD_TYPE
Given the impracticality of mixing Debug and Release binaries on
Windows, we need a way to spot up front when we've been handed a
BRLCAD_EXT_DIR using the wrong MSVC runtime.

Untested (i.e. virtually certain to not be correct yet) but rough in
some logic to try to detect what configuration bext was built in.