wip - update 'alive' checking on linux
kill(pid, 0) is not quite enough for our use cases. If a process
finishes, it will still report alive until it is waited on - which is
not what we want. Instead, attempt to wait on the process and report
that status as its alive state.

Still need to test to make sure we can't double wait (ie infinite hang) on a
process that's been waited for in 'alive' before calling bu_process_wait