shift readParameter logic into Options class
readParameters deals exclusively with Options setting / validation so
add it as a class member instead of handling in main.

refactor option parsing to use BU_OPT table. Some of the options dont
feel quite right yet, so this will allow flexible modifaction should it
become necessary.
would it be worth refactoring 'set' functions to be used as the BU_OPT
callbacks? It'd clean up some of the post option checks..