Do a little bit of cleanup on the distcheck tgts
With bext coming online, there are some implications for distcheck.  An
earlier commit instructed the distcheck builds to re-use the bext
provided by the parent build, and that's an overall BIG win, but in the
(so far theoretical) case where we try to get distcheck working on
Windows, it won't fly due to the Debug/Release incompatibility problem.
When the time comes we'll have to engineer so we can both reuse when
possible and build when necessary, but for the moment we just skip
defining the targets where we can't use them (and where they weren't
known to be working anyway.)

Also remove the bundled libs control variables - we no longer specify
that at the BRL-CAD configure stage.

Definitely more work to do here - post-build cleanup is well and truly
busted by the bext copies, so it looks like the distclean logic didn't
get the memo.  Also need to clean up the pkgIndex and tclIndex files
properly after that logic was reworked.