Rework gqa plot testing into a regression test
These tests will trigger the cache logic, so rework them into a
regression test were we can use BU_DIR_CACHE to localize the cache
activity.

This seems to be the last functional issue for make check to work (at
least on Linux) when disabling the man user cache.  That should imply
that the build and check process is now localized, and we won't have
tests and multiple builds all trying to pile into the global cache at
once.

The way this was tested was 1) removing the .cache/BRL-CAD directory and
replacing it with a file (i.e. ensuring that any attempt to work with
the standard cache location would fail) and 2) altering cache.c in librt
to make any failure to set up the cache immediately fatal.  If we don't
do that, then the uncached behavior is used and everything still works -
but for this case, we wanted specifically to identify cases where we
didn't have localized cache logic, so we needed an immediate failure.