Realign distcheck tests for bext build testing
The enableall and autodetect settings in the original tests aren't
directly relevant for all cases now, but they do have analogs if the
configure process is responsible for building bext rather than re-using
the parent supplied files.

Set up some additional tests to try and cover the configure-driven bext
cases, and explicitly pass BRLCAD_EXT_DIR for the others rather than
hard coding it into the test templates.

Also simply definition of distcheck target.  distcheck-full is what we
use to actually test - the shorter list of configurations originally
defined for distcheck-std is not used in practice.  If any of the
distcheck-full configurations is broken, that identifies an issue we
need to fix - rather than defining a macro and getting fancy, just
specify the configurations and targets we actually use.