Replace poly2tri with detria
This logic is exposed via the e -m1 shaded drawing and brep->bot
facetization routines.  The latter aren't working for most inputs at the
moment with either triangulator (need to fix that, but that logic is a
deep dive so it will have to be later).  However, testing a couple of
basic primitive->brep->bot conversions confirms we ARE able to get
triangles out with the new detria logic in this pipeline when the
inputs aren't problematic for other reasons.

More significant at the moment is the use of this triangulation for
interactive shaded display of NURBS brep geometry, and in testing so far
the speed appears to be on the same order as poly2tri and the results
are either comparable or (in a couple of cases) actually better. A
conversion of the GrabCAD model YachtMod(55).3dm resulted in inputs that
resulted in the poly2tri-backed draw crashing but the detria-backed draw
succeeding.

Give detria appears to *fix* a drawing problem, that's motivation for
switching it in.  Although not decisive in and of itself, another
convenient benefit of this switch will be eliminating the need for our
forked copy of poly2tri (we have diverged from upstream, and the last
attempt at using vanilla upstream poly2tri resulted in unsuccessful
shading results for various NURBS surfaces.)