Adjust how brep_loop_points is allocated
We were getting the following warning in Release mode with gcc on Linux:

src/libbrep/cdt/fast.cpp:2067:97: error: argument 1 value
‘18446744073709551615’ exceeds maximum object size 9223372036854775807
[-Werror=alloc-size-larger-than=]
 2067 |     ON_SimpleArray<BrepTrimPoint> *brep_loop_points = new ON_SimpleArray<BrepTrimPoint>[loop_cnt];
      |
^ /usr/include/c++/13/new:128:26: note: in a call to allocation function
‘operator new []’ declared here
  128 | _GLIBCXX_NODISCARD void* operator new[](std::size_t) _GLIBCXX_THROW (std::bad_alloc)

To avoid this we instead make a pointer array and create the
BrepTrimPoint arrays individually.  In practice loop_cnt probably
wouldn't have gotten big enough to produce an array object that would
hit this limit, but in principle it IS possible (i.e. the data types
don't preclude it.)