Update src/bext submodule to get librt+geogram gchecker fix
It turns out the gchecker failure in Windows Release mode was due to
linking in a geogram built without the win32 building patches.

The symptoms were that gchecker would launch but hang, with bu_parallel
launching its threads but failing to start its workers successfully.
Even commenting out the actual guts of the work function still resulted
in hanging, indicating the problem was with the parallelism itself.
Bisecting identified the commit merging bext usage into the facetize
branch as the point where the regression first appeared, and commenting
out the librt usage of geogram in that commit resulted in success.

dumpbin /dependents geogram.dll on the failing geogram build produced a
very short dependencies list, which may be related.  In any case, it
looks like the latest bext patch does succeed in allowing gchecker to
run again.