Adjust bot repair to accept multiple inputs
The most common use case for this is most likely going to be "I have a
.g file, fix all the BoTs in it."  Update the usability of the bot
repair command accordingly - now "bot repair *" will try to repair all
the BoT solids in a database.  For the case where a user wants to write
the repair of a single bot to an object name, the -o option is now
available rather than using an optional second argument.