lots of archer editing now results in runtime errors.
in preference/value/expediency order, we'll need to map them in ged, revert their ged_exec, or remove the functionality entirely.  seeing as the editing features are valuable progress towards scriptability, probably best to map them as commands and ideally put testing in place that indicates their use in archer.