Fix shaded drawing of BoTs in CW mode
Noticed this working with bot flip - the bot flip code (incorrectly, as
it happens, given the intent there) was changing the orientation of the
bot from RT_BOT_CCW to RT_BOT_CW, but the shaded display was still
showing the bot as inside-out even though the orientation on the
"flipped" geometry was in fact changed.  Dug a bit, and the poly drawing
routine was ignoring the orientation flag when supplying drawing
instructions.