Add a way to determine if a BoT is inside-out
Usually we end up spotting these cases when a shaded solid drawing winds
up black instead of red, even with no color set on the object.  Add a
way to programmatically detect when this is happening, so we can try to
either do something about it automatically or at least report the
situation to the modeler without them having to manually draw and
inspect the solids to spot it.