Add a lint check for bots with thin face thicknesses
Although the Manifold boolean usually succeeds, it sometimes does so
when faced with coplanar faces by producing extremely close together
faces that raytrace with zero thickness but still show up as hits in
nirt and when drawing the geometry.

We test for this case by finding the center point of each triangle,
backing it off slightly from the triangle plane, and shooting in the
opposite direction of the triangle normal into the mesh.  We then check
the first partition distance to see if that face is part of one of these
degenerate volume pairings.  If such a case is found, abort the test and
report a problem case.

The "fix" for this is typically going to be to tweak the parent CSG
solid definitions to avoid the coplanar faces.  It's probable the zero
distance partitions can also be filtered and ignored analytically in the
solid shotlining cases (rtweight, for example, won't add weight from a
zero length segment), but they remain a visual distraction.
Topologically these *are* valid meshes, but if we can it would be good
to find a way to eliminate the coplanarity artifacts.

One possibility that suggests itself, given the new repair capabilities,
would be to analyze all faces on the mesh and flag any with this
characteristic for removal.  We could then apply the bot repair logic to
the mesh with the zero distance faces removed to create a new valid
manifold.  Need to experiment and see if that would work - not entirely
sure if the repair logic could handle the sort of inputs that process
would produce.