Expose the "remove degenerate thickness faces" feature as a "bot strip" subcommand
Even if the default settings aren't able to repair these meshes, it may be that
manual intervention with more aggressive bot repair settings can.  Failing
that, this allows the modeler to remove the problematic faces and prepare a BoT
for export to other tools for additional repair work.  For that reason, no
attempt is made to automatically repair the output as part of the strip process
- if there are any problematic triangles, the resulting bot will be
non-manifold until repaired.

Going ahead and defining this command rather than capturing the intermediate
results from facetize to avoid complexity in the facetize code, as well as
provide more flexibility for users.

The lint command can produce a list of bot objects that will make sense to
consider as possible inputs for this command.  Note that providing non-manifold
inputs to this command will likely remove triangles in unexpected ways since
there is a risk that the rays shotting from the triangles centers will have
nothing on the opposite side of the mesh to hit.