Set up libged lint cmd versions of checks
We're going to want more powerful reporting and visualization than makes
sense for librt.  Reduce the bot check there to just the one facetize
uses, and shift the more extensive logic into lint.  If we want to set
up for things like json logging, visualization of problem triangles,
etc. it makes more sense to do it in the lint command rather than trying
to shoehorn it into the librt API, at least in the early stages.  If
something emerges that makes sense in librt we can move it back later.