Rough in bot plot subcommand
For now this is just plotting faces by index.  Need to think about
whether it makes sense to support plotting vertices as well - that might
be less useful, since if we can retrieve a vertex's coordinates by index
it would be trivial to either move the view center to that point or make
a sphere there...

On a related note - probably need to adjust the bot get syntax. Right
now it's property then objname, but if we want to upgrade it to allow
getting per face and per vertex info we need the property to optionally
have arguments - that would make more sense in a "bot get objname prop
[opt]" layout.