Add a visualize option to lint
We want to be able to see where the problem triangles in a mesh are, and we
have enough information from the raytrace to be able to do so.  Set up a
-V option for lint to allow plotting of problematic faces.

Ideally we probably want the -I option to lint to allow for more specialization
- i.e., only run one specific type of BoT check, for example.  This will need
some sort of flexible test specification mechanism.