Add an option to limit lint reporting based on triangle size
We can't do it for all shotlines - if we miss a small triangle somewhere other
than at the shotline origin, we won't know about it in the callbacks - but add
an option to allow users to avoid reporting issues when the source triangles
have a small area.

In principle this shouldn't ever be necessary, but sufficiently small triangles
are currently a problem for librt's BoT intersection routines and given we're
deliberately sampling *every* triangle we're guaranteed to report all BoTs with
such small triangles as invalid even if a ray hitting such a small face is
highly improbable.

Reporting the errors remains the default behavior, since the mesh in question
really isn't behaving the way the raytracer expects even if it's a raytracing
problem rather than a geometric problem per say, but this option gives users a
way to determine what category of problem they are seeing and decide if the
circumstances warrant using decimating to try and clear the small triangles.

(Note: in testing, bot decimate -f 0.005 was able to produce a valid bot by
clearing a small triangle, but the -e OpenMesh-based option isn't producing
manifold output and doesn't repair successfully... arrgh.)